requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
streamlit
pycaret
sweetviz
langchain
openai
gradio
tensorflow
torch
plotly
python-dotenv
langchain_google_genai
reportlab
langchain_groq
faiss-cpu
pycaret
sentence-transformers
groq